+++ ui/src/pages/products/Products.page.tsx
@@ -1,17 +1,26 @@
-import { FC, Fragment, useCallback, useEffect } from "react";
+import { FC, useEffect } from "react";
 import { useParams } from "react-router-dom";
 // ..images
-import { useAppSelector } from "@/common/state/store";
+import { useAppDispatch, useAppSelector } from "@/common/state/store";
 import { IProductRecord } from "@/common/interfaces";
-import ProductPreviewSection from "../admin/products/sections/Product.preview.section";
+import ProductPreviewSection from "@/pages/admin/products/sections/Product.preview.section";
 import Loader from "@/components/loaders/Loader";
+import useLocalStorage from "@/hooks/use-localstorage";
+import { setCheckoutCount } from "@/common/state/features/checkout/checkout.slice";
 
 const ProductsPage: FC = () => {
-
+  const {storedValue, setStoredValue } = useLocalStorage<IProductRecord[]>('products', []);
   const productState = useAppSelector((state) => state.products);
   const assetState = useAppSelector((state) => state.asset);
   const { id } = useParams();
+  const dispatch = useAppDispatch();
 
+  // evt handlers
+  const addToCartHandler = (item: IProductRecord) => {
+    console.log('item to add to cart: ', item);
+    setStoredValue(prev => [...prev, item]);
+  };
+  
   useEffect(() => {
     if (Boolean(id) && id !== "") {
       document.getElementById(id)?.scrollIntoView({
@@ -21,6 +30,13 @@ const ProductsPage: FC = () => {
       });
     }
   }, [id]);
+
+
+  useEffect(()=>{
+    console.log('Products: Stored item is: ', storedValue);
+    dispatch(setCheckoutCount(storedValue.length));
+  },[ storedValue ])
+
 
   if (
     Boolean(productState.product_list) && 
@@ -32,6 +48,7 @@ const ProductsPage: FC = () => {
       <div className="container relative z-10">
         {productState.product_list.map((item: IProductRecord, idx: number) => (
           <ProductPreviewSection
+            onAddCart={addToCartHandler}
             isAdmin={false}
             key={idx}
             selectedProduct={item}